requirements.txt
# FastAPI & Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlmodel==0.0.14
sqlite3

# Telegram
pyrogram==2.0.106
tgcrypto==1.2.5

# HTTP Client
httpx==0.26.0

# Security
cryptography==42.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Webhooks & APIs
stripe==7.9.0
twilio==8.12.0

# Environment
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
